Changed handling of mesh level transforms, they are now 'experimental'.
The previous scale check for mesh level transforms did not properly
compare the three floating point components of the active object's
scale vector. Users must now check a box to enable experimental
mesh level transforms, and will be informed that non uniform
scaling is not supported. To be removed when non-uniform scaling
is supported.