Modified quick tool usage and made "Mesh Piece" transforms easier.
"Mesh Piece" transforms no longer use/need vertex groups (which
were awkward/not ideal). All quick tools now explicitly show users
that the "source" for a transformation is auto-grabbed from selected
verts, and can disable that option to manually grab the transformation
source, so that target geometry for the transformation can be
selected in the viewport instead.